+++ frontend/src/pages/study-room-page/tabs/notice-tab-panel/NoticeTabPanel.tsx
@@ -1,19 +1,15 @@
-import { Link, Navigate, useLocation, useParams } from 'react-router-dom';
+import { Navigate, useLocation, useParams } from 'react-router-dom';
 
-import { type Theme, useTheme } from '@emotion/react';
+import { useTheme } from '@emotion/react';
 
 import { PATH } from '@constants';
 
-import type { StudyId } from '@custom-types';
-
 import { useUserRole } from '@hooks/useUserRole';
 
-import { TextButton } from '@shared/button';
-import Divider from '@shared/divider/Divider';
-import Flex from '@shared/flex/Flex';
 import PageWrapper from '@shared/page-wrapper/PageWrapper';
 
-import ArticleList from '@notice-tab/components/article-list/ArticleList';
+import ArticleListPage from '@study-room-page/tabs/notice-tab-panel/components/article-list-page/ArticleListPage';
+
 import Article from '@notice-tab/components/article/Article';
 import Edit from '@notice-tab/components/edit/Edit';
 import Publish from '@notice-tab/components/publish/Publish';
@@ -24,7 +20,7 @@ const NoticeTabPanel: React.FC = () => {
   const { studyId: _studyId, articleId: _articleId } = useParams<{ studyId: string; articleId: string }>();
   const [studyId, articleId] = [Number(_studyId), Number(_articleId)];
 
-  const { isOwner, isNonMember, isMember } = useUserRole({ studyId });
+  const { isNonMember, isMember } = useUserRole({ studyId });
 
   const lastPath = location.pathname.split('/').at(-1);
 
@@ -38,34 +34,10 @@ const NoticeTabPanel: React.FC = () => {
           if (lastPath === 'edit') return <Edit studyId={studyId} articleId={articleId} />;
           return <Article studyId={studyId} articleId={articleId} />;
         }
-        return <ArticleListPage theme={theme} studyId={studyId} isOwner={isOwner} />;
+        return <ArticleListPage theme={theme} studyId={studyId} />;
       })()}
     </PageWrapper>
   );
 };
 
 export default NoticeTabPanel;
-
-type ArticleListPageProps = {
-  theme: Theme;
-  studyId: StudyId;
-  isOwner: boolean;
-};
-const ArticleListPage: React.FC<ArticleListPageProps> = ({ theme, studyId }) => {
-  const { isOwner } = useUserRole({ studyId }); // @TODO: 객체로 받을 필요가 있나?
-  return (
-    <>
-      <Flex justifyContent="flex-end">{isOwner && <PublishPageLink />}</Flex>
-      <Divider color={theme.colors.secondary.dark} space="8px" />
-      <ArticleList studyId={studyId} />
-    </>
-  );
-};
-
-const PublishPageLink: React.FC = () => (
-  <Link to={PATH.NOTICE_PUBLISH}>
-    <TextButton variant="primary" custom={{ fontSize: 'lg' }}>
-      글쓰기
-    </TextButton>
-  </Link>
-);
